requirements.txt
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
networkx==3.2.1
google-generativeai>=0.8.0
google-genai>=1.45.0
python-dotenv==1.0.0
plotly
matplotlib